Remove useless files (old analysis, dataframe and core analysis) and change the name of analysis_duplicate to analysis
