--- src/core/__init__.py
@@ -1,7 +1,0 @@
-from .agregation import *  # noqa: F401, F403
-from .analysis import *  # noqa: F401, F403
-from .dataframe import *  # noqa: F401, F403
-from .interpretation_functions import *  # noqa: F401, F403
-from .measures_functions import *  # noqa: F401, F403
-from .schemas import *  # noqa: F401, F403
-from .weighting import *  # noqa: F401, F403
