+++ src/resources/__init__.py
@@ -1,2 +1,2 @@
-from .analysis_duplicate import *  # noqa: F401, F403
+from .analysis import *  # noqa: F401, F403
 from .available import *  # noqa: F401, F403
